+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Current
 
-- Separate phantom argument from the definitions of `fset`, `fmap` and `fperm`
+- Separate phantom argument from the definitions of `fset`, `fmap` and `fperm`.
 
 - Add `ordType` instances for `mathcomp.choice.GenTree.tree` and `tuple`.
 
@@ -8,6 +8,8 @@
   `dommPn`, `codommP` and `codommPn`.
 
 - Miscellaneous lemmas for finite sets.
+
+- Version of `mapm` that allows modifying the domain of a map.
 
 # 0.1.0 (2018/04/26)